attempt to fix dev modes
