+++ packages/frontend/src/features/Navigation/TopBar.tsx
@@ -1,5 +1,4 @@
 import React, { ReactElement } from 'react';
-import { Icon } from '@iconify/react';
 import { Divider } from '@mantine/core';
 import { mergeDefaultTailwindClassnames } from '../../utils/mergeDefaultTailwindClassnames';
 import LoginButton from '../../components/Login/LoginButton';
@@ -7,76 +6,7 @@ import LoginAccountButton from '../../components/Login/LoginAccountButton';
 import { extractClassName } from './utils';
 import { LoginButtonVisibility } from '../../components/Login/types';
 import { StylingOverrideWithMergeControl } from '../../types';
-import { IconSize } from '../DataLibrary/types';
-
-export interface NameAndIcon {
-  readonly name: string;
-  readonly iconSize?: string;
-  readonly rightIcon?: string;
-  readonly leftIcon?: string;
-  readonly classNames?: StylingOverrideWithMergeControl;
-  readonly drawBorder?: boolean;
-}
-
-export interface TopIconButtonProps extends NameAndIcon {
-  readonly href: string;
-  readonly tooltip?: string;
-}
-
-const TopIconButton = ({
-  name,
-  leftIcon = undefined,
-  rightIcon = undefined,
-  iconSize = 'md',
-  classNames = {},
-  drawBorder = true,
-}: NameAndIcon) => {
-  const classNamesDefaults = {
-    root: `flex items-center align-middle px-2 my-2`,
-    button:
-      'flex flex-nowrap items-center align-middle border-b-2 hover:border-accent border-transparent',
-    leftIcon: 'text-secondary-contrast-lighter pr-1',
-    label: 'font-content text-secondary-contrast-lighter block',
-    rightIcon: 'text-secondary-contrast-lighter pl-1',
-  };
-  const mergedClassnames = mergeDefaultTailwindClassnames(
-    classNamesDefaults,
-    classNames,
-  );
-
-  // get the icon size otherwise use the value of iconsSize as a string value: e.g. 2em
-  const iconSz = IconSize[iconSize] ?? iconSize;
-
-  return (
-    <div
-      className={extractClassName('root', mergedClassnames)}
-      aria-label={name}
-    >
-      <div
-        className={extractClassName('button', mergedClassnames)}
-        role="button"
-      >
-        {leftIcon ? (
-          <Icon
-            icon={leftIcon}
-            width={iconSz}
-            height={iconSz}
-            className={extractClassName('leftIcon', mergedClassnames)}
-          />
-        ) : null}
-        <p className={extractClassName('label', mergedClassnames)}> {name} </p>
-        {rightIcon && rightIcon.length > 0 ? (
-          <Icon
-            width={iconSz}
-            height={iconSz}
-            icon={rightIcon}
-            className={extractClassName('rightIcon', mergedClassnames)}
-          />
-        ) : null}
-      </div>
-    </div>
-  );
-};
+import { TopIconButton, TopIconButtonProps } from './TopIconButton';
 
 const processTopBarItems = (
   items: TopIconButtonProps[],
@@ -99,7 +29,6 @@ const processTopBarItems = (
             leftIcon={item.leftIcon}
             rightIcon={item.rightIcon}
             classNames={mergedClassnames}
-            drawBorder={needsBorder}
           />{' '}
           <Divider
             size="md"
@@ -169,7 +98,7 @@ const TopBar = ({
             extractClassName('divider', mergedClassnames),
           )}
           {loginButtonVisibility != LoginButtonVisibility.Hidden ? (
-            <div className="flex items-center">
+            <div className="flex items-center [&>*:only-child]:hidden">
               <LoginAccountButton
                 className={extractClassName('login', mergedClassnames)}
               />
