+++ packages/sampleCommons/config/icons/color.json
@@ -1,6 +1,6 @@
 {
 	"prefix": "color",
-	"lastModified": 1738707650,
+	"lastModified": 1742604489,
 	"icons": {
 		"analysis": {
 			"body": "<g fill=\"currentColor\"><path d=\"m52 22 3.293 3.293L44 36.586l-12-12-12.707 12.707 1.414 1.414L32 27.414l12 12 12.707-12.707L60 30v-8z\"/><path d=\"M20 42a4 4 0 110-8 4 4 0 010 8m0 6a2 2 0 11-.001-3.999A2 2 0 0120 48m12-8a2 2 0 11-.001-3.999A2 2 0 0132 40m12-10a2 2 0 11-.001-3.999A2 2 0 0144 30M20 16a2 2 0 11-.001-3.999A2 2 0 0120 16m0 14a2 2 0 11-.001-3.999A2 2 0 0120 30m16-14a2 2 0 11-.001-3.999A2 2 0 0136 16m-10 6a2 2 0 11-.001-3.999A2 2 0 0126 22m12 26a2 2 0 11-.001-3.999A2 2 0 0138 48m12-2a2 2 0 11-.001-3.999A2 2 0 0150 46M32 30a4 4 0 110-8 4 4 0 010 8m12 12a4 4 0 110-8 4 4 0 010 8\"/><path d=\"M18 38a3.99 3.99 0 013-3.859A4 4 0 0020 34a4 4 0 000 8c.347 0 .679-.058 1-.141-1.722-.447-3-1.997-3-3.859m12-12a3.99 3.99 0 013-3.859A4 4 0 0032 22a4 4 0 000 8c.347 0 .679-.058 1-.141-1.722-.447-3-1.997-3-3.859m12 12a3.99 3.99 0 013-3.859A4 4 0 0044 34a4 4 0 000 8c.347 0 .679-.058 1-.141-1.722-.447-3-1.997-3-3.859\"/><path d=\"m60 54-6-6v4H12V14h4l-6-6-6 6h4v42h46v4z\"/><path d=\"M10 14H8v42h2v-4zm0-6-6 6h2l5-5zm42 12a8 8 0 110-16 8 8 0 010 16\"/><path d=\"M46 12c0-4.079 3.055-7.438 7-7.931A8 8 0 0052 4a8 8 0 000 16c.339 0 .672-.028 1-.069-3.945-.493-7-3.852-7-7.931\"/><path d=\"m52 14.707-2-2-1.646 1.647-.708-.708L50 11.293l2 2 1.646-1.647.708.708z\"/><path d=\"m52 10 4 4v-4z\"/></g>",
