+++ packages/sampleCommons/config/icons/dataDictionary.json
@@ -1,6 +1,6 @@
 {
 	"prefix": "dataDictionary",
-	"lastModified": 1738707650,
+	"lastModified": 1742604489,
 	"icons": {
 		"dictionary-icon-administrative": {
 			"body": "<g fill=\"currentColor\" fill-rule=\"evenodd\"><path d=\"M5.5 4.9h.72c.013.005.026.015.038.016.261.015.522.023.782.044.466.037.93.101 1.385.253.377.126.73.322 1.018.707.038.05.057.1.057.177q-.002 3.847-.001 7.696v.107c-.018-.007-.025-.008-.031-.012q-.118-.082-.234-.17a3.8 3.8 0 00-1.973-.803c-.51-.053-1.023-.07-1.534-.101-.07-.005-.142-.001-.22-.001V5.036c0-.046-.005-.09-.007-.136m8.28 0h.72q-.005.067-.007.135v7.63l-.004.136h-.077a13 13 0 00-1.784.118 4 4 0 00-1.863.79c-.084.064-.17.123-.265.191v-.107q0-3.854.002-7.707c0-.051.02-.115.045-.15q.222-.309.509-.484c.33-.205.68-.304 1.035-.375.506-.1 1.015-.14 1.526-.158.054-.002.108-.013.162-.019\"/><path d=\"M3.5 5.9h.983v8.114h.12c.61.012 1.22.01 1.828.039.717.033 1.434.105 2.121.374.316.123.618.295.925.446.01.004.016.018.023.027h-6zm8 9c.006-.01.011-.022.019-.028.5-.36 1.027-.598 1.584-.702a10 10 0 011.686-.148l.799-.008h.093V5.9h.819v9z\"/></g>"
